+++ source/extensions/lightspeed.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.10.1]
+### Added
+- Added constants for ComfyUI window name
+
 ## [1.10.0]
 ### Added
 - Added REMIX_LOGIC_GRAPH window name constant