+++ source/extensions/lightspeed.trex.comfyui.widget/docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Added
+- Added ability to add selected textures and meshes to the ComfyUI queue
+- Added a window to display the ComfyUI UI
+
+### Changed
+- Improved the layout for a more production-ready UI
+
 ## [1.0.0]
 ### Added
 - Created